# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Potoo API
  version: "0.1.0"
  description: |
    Developer-first notification orchestration platform.

    Authenticate all requests with an API key in the `Authorization` header:

    ```
    Authorization: Bearer <api_key>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API key issued from the Potoo dashboard.
  schemas:
    AuthUser:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    OrgSummary:
      type: object
      required: [id, name, slug, type, role, is_active]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [platform, customer]
        role:
          type: string
          enum: [owner, member]
        is_active:
          type: boolean
        api_key_count:
          type: integer
          description: Number of active (non-revoked) API keys — included in platform org listing
        call_count:
          type: integer
          description: Total notifications sent — included in platform org listing
    UpdateOrgRequest:
      type: object
      properties:
        is_active:
          type: boolean
...
    SelectOrgRequest:
      type: object
      required: [org_id]
      properties:
        org_id:
          type: string
    CreateOrgRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          example: Acme Corp
        slug:
          type: string
          example: acme-corp
...
    CreateInviteRequest:
      type: object
      required: [org_id]
      properties:
        org_id:
          type: string
    InviteResponse:
      type: object
      required: [id, org_id, token, expires_at, created_at]
      properties:
        id:
          type: string
        org_id:
          type: string
        org_name:
          type: string
        token:
          type: string
        used_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    InvitePreview:
      type: object
      required: [org_name, expires_at]
      properties:
        org_name:
          type: string
        expires_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: "Template is missing required variable: first_name"
            request_id:
              type: string
              example: req_01HX1234
            docs_url:
              type: string
              example: https://docs.potoo.com/errors/validation_failed
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/invite/{token}:
    get:
      operationId: getInvite
      summary: Validate an invite token and return org info for the invite page
      tags: [auth]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valid invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invite already used or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/auth/select-org:
    post:
      operationId: selectOrg
      summary: Switch active org — deletes current session and starts a new one
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectOrgRequest"
      responses:
        "200":
          description: Org selected — new session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not a member of the requested org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me/orgs:
    get:
      operationId: listMyOrgs
      summary: List all orgs the current user belongs to
      tags: [auth]
      security: []
      responses:
        "200":
          description: Org list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/orgs:
    get:
      operationId: listOrgs
      summary: List all customer organizations (platform owner only)
      tags: [platform]
      responses:
        "200":
          description: Org list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgSummary"
    post:
      operationId: createOrg
      summary: Create a customer organization (platform owner only)
      tags: [platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Org created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/orgs/{orgId}:
    patch:
      operationId: updateOrg
      summary: Update a customer organization (platform owner only)
      tags: [platform]
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
      responses:
        "200":
          description: Org updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"
        "404":
          description: Org not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/invites:
    post:
      operationId: createInvite
      summary: Create a single-use invite link for a customer org (platform owner only)
      tags: [platform]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
        "404":
          description: Org not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listInvites
      summary: List all invites (platform owner only)
      tags: [platform]
      security: []
      responses:
        "200":
          description: Invite list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/InviteResponse"
...